added a fallback integratino test
